+++ resources/i18n/app_en_US.ts
@@ -5605,8 +5605,8 @@ This operation will:
         <location filename="../../build/msvc2022/Release/EnhanceVision/qml/pages/SettingsPage.qml" line="1869"/>
         <location filename="../../qml/pages/SettingsPage.qml" line="1869"/>
         <location filename="../../qml/pages/SettingsPage.qml" line="1869"/>
-        <source>Built with Qt6 + NCNN</source>
-        <translation>Built with Qt6 + NCNN</translation>
+        <source>Implemented based on technologies such as Qt6 + NCNN + Real - ESRGAN...</source>
+        <translation>Implemented based on technologies such as Qt6 + NCNN + Real - ESRGAN...</translation>
     </message>
     <message>
         <location filename="../../build/msvc2022/Debug/EnhanceVision/qml/pages/SettingsPage.qml" line="1246"/>